Update GentleProstate_Analysis.lua
This code for analysis now works. We are able to compare our auto-contour and a doctors contour. 

Note that the number 50 in the function AVS:FIELD_ISOSURFACE(wm.scan[4].data, refDots, refIndex, 50), refers to half of the burn value. I.e. including pixels on the outskirts at half the burn intensity to be included.