Patients for atlas generator
This is a piece of code, originally written by Tom with edits made by either of us. This code uses volume information for the body, prostate and rectum, splits this into percentiles, then randomly selects a patient from each to be used in our ATLAS for auto-contouring. The purpose of this code is to introduce variation within our ATLAS.