Update PatientsForAtlasGenerator.py
I have changed the way in which we are creating subsets. I have changed this such that each patient is selected closer to the percentile, i.e. for patients about the 50th percentile, randomly select patients between the 49th-51st percentile, instead of the 50th-55th. Doing so ensure we are not always 'picking to the right'.

I had to change random_integers() -> randint() as this was yielding a range error